# LEAF-4770 - Optimization: Use userMetadata instead of individual lookups